+++ tests/testthat/test-QTableToBasicTable.R
@@ -1,4 +1,4 @@
-context("QTable to BasicTable")
+context("QTable to TidyTabularData")
 
 ## Load QTables for testing
 ## Need to specify use of base's system.file not devtools
@@ -19,7 +19,7 @@ test_that("isQTable works",
 test_that("5D case",
 {
     expect_true(isQTable(q5))
-    expect_warning(out <- AsBasicTable(q5), "\\bAverage\\b", perl = TRUE)
+    expect_warning(out <- AsTidyTabularData(q5), "\\bAverage\\b", perl = TRUE)
     dims <- dim(q5)
     dn.orig <- dimnames(q5)
     dn.new <- dimnames(out)
@@ -32,7 +32,7 @@ test_that("5D case",
 test_that("4D table, pick any grid x multi cat with multiple statistics",
 {
     expect_true(isQTable(q4.ms))
-    expect_warning(out <- AsBasicTable(q4.ms), "Column [%]")
+    expect_warning(out <- AsTidyTabularData(q4.ms), "Column [%]")
     dims <- dim(q4.ms)
     expect_equal(dim(out), c(dims[1L]*dims[3L], dims[2L]))
     expect_equal(out["Diet Coke: NET", "Traditional"], q4.ms["Diet Coke", "Traditional", "NET", "Column %"])
@@ -58,7 +58,7 @@ test_that("Flattening 4D table",
 test_that("4D Qtable numeric grid x cat. grid only one statistic",
 {
     expect_true(isQTable(q4.os))
-    expect_silent(out <- AsBasicTable(q4.os))
+    expect_silent(out <- AsTidyTabularData(q4.os))
     expect_equal(attr(out, "name"), attr(q4.os, "name"))
     expect_equal(attr(out, "statistic"), attr(q4.os, "statistic"))
 })
@@ -66,51 +66,51 @@ test_that("4D Qtable numeric grid x cat. grid only one statistic",
 test_that("3D number grid by number multi, one statistic",
 {
     expect_true(isQTable(q3.os))
-    expect_silent(out <- AsBasicTable(q3.os))
+    expect_silent(out <- AsTidyTabularData(q3.os))
 })
 
 test_that("3D: categorical by pick1 multi, one statistic",
 {
     expect_true(isQTable(q3.os2))
-    expect_silent(out <- AsBasicTable(q3.os2))
+    expect_silent(out <- AsTidyTabularData(q3.os2))
     expect_length(dim(out), 2L)
 })
 
 test_that("3D: categorical x numeric, multi stat.",
 {
     expect_true(isQTable(q3.ms))
-    expect_warning(out <- AsBasicTable(q3.ms))
+    expect_warning(out <- AsTidyTabularData(q3.ms))
     expect_length(dim(out), 2L)
 })
 
 test_that("3D: grid, multi stat",
 {
     expect_true(isQTable(q3.ms2))
-    expect_warning(out <- AsBasicTable(q3.ms2))
+    expect_warning(out <- AsTidyTabularData(q3.ms2))
 })
 
 test_that("3D: multi stat. character entries",
 {
-    expect_warning(out <- AsBasicTable(q3.ms3))
+    expect_warning(out <- AsTidyTabularData(q3.ms3))
     expect_length(dim(out), 2L)
 
     ## first stat. can't be converted to numeric
     x <- q3.ms3
     x[,, 1] <- x[,, 2]
-    expect_warning(out <- AsBasicTable(x), regexp = "QTable contains character entries",
+    expect_warning(out <- AsTidyTabularData(x), regexp = "QTable contains character entries",
                    all = FALSE)
     expect_is(out[1, 1], "character")
 
     ## empty chars are converted to NA with no warning
     x <- q3.ms3
     x[1:3, 1, 1] <- ""
-    expect_warning(out <- AsBasicTable(x), "^Multiple statistics", all = TRUE)
+    expect_warning(out <- AsTidyTabularData(x), "^Multiple statistics", all = TRUE)
     expect_equal(sum(is.na(out)), 3L)
 
     ## single statistic table with characters
     x <- q3.ms3
     attr(x, "statistic") <- "mean"
-    expect_warning(out <- AsBasicTable(x), regexp = "QTable contains character entries",
+    expect_warning(out <- AsTidyTabularData(x), regexp = "QTable contains character entries",
                    all = FALSE)
     expect_equal(dim(out), c(dim(q3.ms3)[1]*dim(q3.ms3)[3], dim(q3.ms3)[2]))
 })
@@ -118,39 +118,39 @@ test_that("3D: multi stat. character entries",
 test_that("2D: qdate x cat, 1 stat",
 {
     expect_true(isQTable(q2.os))
-    expect_silent(out <- AsBasicTable(q2.os))
+    expect_silent(out <- AsTidyTabularData(q2.os))
 })
 
 
 test_that("2D: multi grid, 1 stat",
 {
     expect_true(isQTable(q2.os2))
-    expect_silent(out <- AsBasicTable(q2.os2))
+    expect_silent(out <- AsTidyTabularData(q2.os2))
 })
 
 test_that("2D: numeric x cat, 1 stat",
 {
     expect_true(isQTable(q2.os3))
-    expect_silent(out <- AsBasicTable(q2.os3))
+    expect_silent(out <- AsTidyTabularData(q2.os3))
 
 })
 
 test_that("2D: multi cat, multi stat",
 {
     expect_true(isQTable(q2.ms))
-    expect_warning(out <- AsBasicTable(q2.ms))
+    expect_warning(out <- AsTidyTabularData(q2.ms))
 })
 
 test_that("1D: number multi",
 {
     expect_true(isQTable(q1.os))
-    expect_silent(out <- AsBasicTable(q1.os))
+    expect_silent(out <- AsTidyTabularData(q1.os))
 })
 
 test_that("1D: number",
 {
     expect_true(isQTable(q1.os2))
-    expect_silent(out <- AsBasicTable(q1.os2))
+    expect_silent(out <- AsTidyTabularData(q1.os2))
 })
 
 test_that("1D: setDimNames is okay with 1D array",
@@ -162,6 +162,6 @@ test_that("1D: setDimNames is okay with 1D array",
                         questions = c("Preferred cola",
                                       "SUMMARY"))
     expect_true(isQTable(q1))
-    expect_silent(out <- AsBasicTable(q1))
+    expect_silent(out <- AsTidyTabularData(q1))
 })